Recursive delete

+++ src/web3/rm.js
@@ -1,7 +1,16 @@
-import {basename} from 'path'
-import {utf8ToHex} from 'web3-utils'
+import * as Path from 'path'
+import {utf8ToHex, hexToUtf8} from 'web3-utils'
 
 export default async function rm(kernel, path, setProgress, setProgressText) {
-  setProgressText(`Deleting ${basename(path)}`)
-  await kernel.unlink(utf8ToHex(path))
+  const {fileType, entries} = await kernel.stat(utf8ToHex(path))
+  if (fileType != 2) {
+    setProgressText(`Deleting file ${path}`)
+    return await kernel.unlink(utf8ToHex(path))
+  }
+  for (let i = entries-1; i > 1; i--) {
+    const name = hexToUtf8(await kernel.readkeyPath(utf8ToHex(path), i))
+    await rm(kernel, Path.join(path, name), setProgress, setProgressText)
+  }
+  setProgressText(`Deleting directory ${path}`)
+  await kernel.rmdir(utf8ToHex(path))
 }
